% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARX_functions.R
\name{companion.form}
\alias{companion.form}
\title{Companion form function}
\usage{
companion.form(pol)
}
\arguments{
\item{pol}{Coefficient vector. If polynomial is 1 - ax - bx^2, coefficient vector is c(a, b).}
}
\value{
\item{C}{Companion matrix C.}
}
\description{
This function allows you to compute a companion form matrix in order to check the stability of causal and noncausal part of the ARX model.
}
\examples{
pol <- c(0.3,0.4)
C <- companion.form(pol)
}
\author{
Sean Telg
}
\keyword{stability,}
\keyword{stationarity}
