% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARX_functions.R
\name{compute.MA}
\alias{compute.MA}
\title{Coefficients of the moving average representation function}
\usage{
compute.MA(pol, M)
}
\arguments{
\item{pol}{Coefficient vector. If polynomial is 1 - ax - bx^2, coefficient vector is c(a, b).}

\item{M}{Truncation value M (how many MA coefficients should be computed?).}
}
\value{
\item{psi}{Vector containing coefficients of the moving average representation.}
}
\description{
This function allows you to invert a polynomial (either the causal or the noncausal one) and output the corresponding coefficients of the moving average representation.
}
\examples{
pol <- c(0.3,0.4)
psi <- companion.form(pol)
}
\author{
Sean Telg
}
\keyword{stability,}
\keyword{stationarity}
