% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_healthy_subspace.R
\name{marx_find_healthy_subspace}
\alias{marx_find_healthy_subspace}
\title{Find a low-dimensional subspace representing an approximate basis for
healthy single-cell data using matrix factorization}
\usage{
marx_find_healthy_subspace(healthy_data, ..., method = c("svd", "nmf", "pma"))
}
\arguments{
\item{healthy_data}{A data.frame or matrix of dimension (n x p) in which rows
represent cells and columns represent single-cell measurements
(i.e. CyTOF antigens).
Only cells taken from healthy (or reference) samples should be included.}

\item{...}{Additional arguments to pass to the specific method.}

\item{method}{A character vector of length 1 indicating which matrix
factorization method should be used. Valid options are "svd", "nmf", and "pma".}
}
\value{
A \code{marx_subspace} S3 object with 4 slots.
\code{metacells} is a (p x k) matrix in which the
columns represent the k metacells and the p rows represent the p features in the
input data. Each entry represents the weighting of a feature on each metacell.
\code{embeddings} is a (n x k) matrix giving the projection of each of the n
input cells onto the k metacells.
\code{num_metacells} is an integer of length 1 indicating how many metacells
were requested for the subspace representation.
\code{method} is a character vector of length 1 indicating which matrix factorization
method was used.
}
\description{
Find a low-dimensional subspace representing an approximate basis for
healthy single-cell data using matrix factorization
}
