% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_healthy_subspace.R
\name{marx_find_healthy_subspace_nmf}
\alias{marx_find_healthy_subspace_nmf}
\title{Find a low-dimensional subspace representing an approximate basis for
healthy cytometry data using non-negative matrix factorization (NMF).}
\usage{
marx_find_healthy_subspace_nmf(healthy_data, num_metacells = 10L)
}
\arguments{
\item{healthy_data}{A data.frame or matrix of dimension (n x p) in which rows
represent cells and columns represent single-cell measurements
(i.e. CyTOF antigens).
Only cells taken from healthy (or reference) samples should be included.}

\item{num_metacells}{The number of metacells (i.e. the number of NMF
components) to use in constructing the healthy subspace.}
}
\value{
A \code{marx_subspace} S3 object with 4 slots.
\code{metacells} is a (p x k) matrix in which the
columns represent the k metacells and the p rows represent the p features in the
input data. Each entry represents the weighting of a feature on each metacell.
\code{embeddings} is a (n x k) matrix giving the projection of each of the n
input cells onto the k metacells (these values are often called NMF embeddings,
i.e. the first column contains the scores for each of the n input
cells on the first NMF factor).
\code{num_metacells} is an integer of length 1 indicating how many metacells
were specified in the original function call.
\code{method} is a character vector of length 1 indicating which matrix factorization
method was used (in this case, "nmf").
}
\description{
Find a low-dimensional subspace representing an approximate basis for
healthy cytometry data using non-negative matrix factorization (NMF).
}
