% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_healthy_subspace.R
\name{marx_find_healthy_subspace_svd}
\alias{marx_find_healthy_subspace_svd}
\title{Find a low-dimensional subspace representing an approximate basis for
healthy cytometry data using the SVD}
\usage{
marx_find_healthy_subspace_svd(
  healthy_data,
  pca_threshold = 0.95,
  return_recipe = FALSE
)
}
\arguments{
\item{healthy_data}{A data.frame or matrix of dimension (n x p) in which rows
represent cells and columns represent single-cell measurements
(i.e. CyTOF antigens).
Only cells taken from healthy (or reference) samples should be included.}

\item{pca_threshold}{a value between 0 and 1 representing the percentage of
variance in the original data that should be preserved during PCA transformation.
This will determine the number k of PCs (i.e. metacells) in the result.}

\item{return_recipe}{A boolean value indicating if the recipe for the pca
should be returned (TRUE) (or if the projection itself should be returned; FALSE; the
default). Not currently used.}
}
\value{
A \code{marx_subspace} S3 object with 4 slots.
\code{metacells} is a (p x k) matrix in which the
columns represent the k metacells and the p rows represent the p features in the
input data. Each entry represents the weighting of a feature on each metacell.
\code{embeddings} is a (n x k) matrix giving the projection of each of the n
input cells onto the k metacells (these values are often called the principal
components, i.e. the first column contains the scores for each of the n input
cells on the first principal component).
\code{num_metacells} is an integer of length 1 indicating how many PCs/metacells
are needed to preserve the proportion of variance specified by \code{pca_threshold}.
\code{method} is a character vector of length 1 indicating which matrix factorization
method was used (in this case, "svd").
}
\description{
Find a low-dimensional subspace representing an approximate basis for
healthy cytometry data using the SVD
}
