% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marx_utils.R
\name{marx_find_magnitudes}
\alias{marx_find_magnitudes}
\title{Calculate the magnitudes of the disease-specific and lineage-specific
components in a marx projection.}
\usage{
marx_find_magnitudes(marx_projection)
}
\arguments{
\item{marx_projection}{The output data.frame of marx_project()}
}
\value{
A data.frame containing the healthy- and disease-specific component
magnitudes in marx_projection.
}
\description{
Calculate the magnitudes of the disease-specific and lineage-specific
components in a marx projection.
}
