% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_healthy_subspace.R
\name{marx_project}
\alias{marx_project}
\title{Project single-cell measurements onto a PCA-defined healthy subspace}
\usage{
marx_project(new_data, healthy_subspace)
}
\arguments{
\item{new_data}{A tibble in which rows represent cells and columns represent
antigens measured for each cell. Only the columns in \code{new_data} corresponding
to antigens used to build the healthy subspace will be used in the projection.
All others will be ignored.}

\item{healthy_subspace}{A matrix output of the function \code{marx_find_healthy_subspace}}
}
\value{
A tibble with an equal number of rows as \code{new_data} and twice the number
of columns. Each row will represent a cell from \code{new_data}, and each column
will represent either the healthy or disease-specific component of the
healthy subspace projection.

NEEDS TO CHECK THAT THE HEALTHY BASIS IS COMPATIBLE WITH THE new_data IF
CLUSTER_COL IS SPECIFIED
}
\description{
Project single-cell measurements onto a PCA-defined healthy subspace
}
