% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marx_s3_objects.R
\name{new_marx_subspace}
\alias{new_marx_subspace}
\title{TODO?}
\usage{
new_marx_subspace(metacells, embeddings, num_metacells, method)
}
\arguments{
\item{metacells}{TODO?}

\item{embeddings}{TODO?}

\item{num_metacells}{TODO?}

\item{method}{TODO?}
}
\value{
TODO?
}
\description{
TODO?
}
