% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marx_s3_objects.R
\name{print.marx_subspace}
\alias{print.marx_subspace}
\title{Print a marx subspace.}
\usage{
\method{print}{marx_subspace}(x, ...)
}
\arguments{
\item{x}{TODO?}

\item{...}{TODO}
}
\value{
TODO?
}
\description{
Print a marx subspace.
}
