% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MARX_functions.R
\name{regressor.matrix}
\alias{regressor.matrix}
\title{The regressor matrix function}
\usage{
regressor.matrix(y, x, p)
}
\arguments{
\item{y}{Data vector of time series observations.}

\item{x}{Matrix of data (every column represents one time series). Specify NULL or "not" if not wanted.}

\item{p}{Number of autoregressive terms to be included.}
}
\value{
\item{Z}{Regressor matrix}
}
\description{
This function allows you to create a regressor matrix.
}
\examples{
data <- sim.marx(c('t',3,1),c('t',1,1),100,0.5,0.4,0.3)
regressor.matrix(data$y, data$x, 2)
}
\author{
Sean Telg
}
\keyword{estimation}
